Moved bot to database, major refactoring
In this commit, there were major changes to the codebase.

The main one was the replacement of the many smaller test files used for
persistence into a unified SQLite database. PonyORM was chosen as the
ORM for handling of the transactions. This change allowed centralized
acccess to the data, hopefully better lookup times, and reduced
brittleness.

The second major change was that now signatures are dynamically generated,
negating the need to have them stored as images in separate files. I have
no idea why they were placed like this in the first place, seeing that
the script that generated them was easy to include in the main routine.

Finally, general housekeeping happened. Configuration was moved to its
own file.